# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Interacción con Base de Datos FireBird",
        "description": "API para la gestión",
        "version": "1.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtener todos los usuarios",
                "operationId": "e9d4d8b419dbef04b5748fbf9ccf3881",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "4e9dbcbdaaaf9faab196a82d5cbaea23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    }
                }
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "e9ac05ddc4ea46d8848b012d7d92e3ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "27ac608b6dd8b8e7c7c45b10c5210b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "ea700764859a665afd38bb39c2842f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado"
                    }
                }
            }
        },
...
        "/api/usuarios/restore/{id}": {
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Restaurar un usuario eliminado",
                "operationId": "c6e9c297450620d3d3ba5f4482e93808",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario restaurado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Usuario": {
                "title": "Usuario",
                "description": "Modelo de Usuario",
                "required": [
                    "ID_PERSONAS",
                    "LOGIN",
                    "PASSWORD"
                ],
                "properties": {
                    "ID_PERSONAS": {
                        "description": "ID de la persona asociada",
                        "type": "integer"
                    },
                    "LOGIN": {
                        "description": "Login del usuario",
                        "type": "string"
                    },
                    "PASSWORD": {
                        "description": "Contraseña del usuario",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}